% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trustpilot.R
\name{factorToCharacter}
\alias{factorToCharacter}
\title{A trimming function}
\usage{
factorToCharacter(df)
}
\arguments{
\item{df}{The data frame that should be converted}
}
\description{
This function will change a data frame's columns from factor to character
}
\examples{
factorToCharacter()
}
