% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trustpilot.R
\name{trim}
\alias{trim}
\title{A trimming function}
\usage{
trim(str)
}
\arguments{
\item{str}{The string that should be trimmed}
}
\description{
This function will remove white space vefore and after a string
}
\examples{
trim()
}
