% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trustpilot.R
\name{trustpilot}
\alias{trustpilot}
\title{Collecting data from Trustpilot}
\usage{
trustpilot(domain)
}
\arguments{
\item{domain}{The domain name that will be used}
}
\description{
This function will collect review data from Trustpilot
}
\examples{
trustpilot(domain = domain.com)
}
